Extract multilingual model logic into a trait
